+++ R/expand.R
@@ -1,92 +1,80 @@
-#' Expand data frame to include all combinations of values
+#' Expand data frame to include all possible combinations of values
 #'
-#' `expand()` is often useful in conjunction with `left_join()` if
-#' you want to convert implicit missing values to explicit missing values.
-#' Or you can use it in conjunction with `anti_join()` to figure
-#' out which combinations are missing.
+#' @description
+#' `expand()` generates all combination of variables found in a dataset.
+#' It is paired with `nesting()` and `crossing()` helpers. `crossing()`
+#' is a wrapper around [expand_grid()] that de-duplicates and sorts its inputs;
+#' `nesting()` is a helper that only finds combinations already present in the
+#' data.
 #'
-#' `crossing()` is a wrapper around [expand_grid()] that deduplicates and sorts
-#' each input. `nesting()` is the complement to `crossing()`: it only keeps
-#' combinations of values that appear in the data.
+#' `expand()` is often useful in conjunction with joins:
+#'
+#'  * use it with `right_join()` to convert implicit missing values to
+#'    explicit missing values (e.g., fill in gaps in your data frame).
+#'  * use it with `anti_join()` to figure out which combinations are missing
+#'    (e.g., identify gaps in your data frame).
 #'
 #' @inheritParams expand_grid
 #' @param data A data frame.
 #' @param ... Specification of columns to expand. Columns can be atomic vectors
 #'   or lists.
 #'
-#'   To find all unique combinations of x, y and z, including those not
-#'   found in the data, supply each variable as a separate argument.
-#'   To find only the combinations that occur in the data, use nest:
-#'   `expand(df, nesting(x, y, z))`.
+#'   * To find all unique combinations of `x`, `y` and `z`, including those not
+#'     present in the data, supply each variable as a separate argument:
+#'     `expand(df, x, y, z)`.
+#'   * To find only the combinations that occur in the
+#'     data, use `nesting`: `expand(df, nesting(x, y, z))`.
+#'   * You can combine the two forms. For example,
+#'     `expand(df, nesting(school_id, student_id), date)` would produce
+#'     a row for each present school-student combination for all possible
+#'     dates.
 #'
-#'   You can combine the two forms. For example,
-#'   `expand(df, nesting(school_id, student_id), date)` would produce
-#'   a row for every student for each date.
+#'   When used with factors, `expand()` uses the full set of levels, not just
+#'   those that appear in the data. If you want to use only the values seen in
+#'   the data, use `forcats::fct_drop()`.
 #'
-#'   For factors, the full set of levels (not just those that appear in the
-#'   data) are used. For continuous variables, you may need to fill in values
-#'   that don't appear in the data: to do so use expressions like
+#'   When used with continuous variables, you may need to fill in values
+#'   that do not appear in the data: to do so use expressions like
+
 #'   `year = 2010:2020` or `year = \link{full_seq}(year,1)`.
-#'
-#'   Length-zero (empty) elements are automatically dropped.
-#' @seealso [complete()] for a common application of `expand`:
-#'   completing a data frame with missing combinations. [expand_grid()]
-#'   is low-level that doesn't deduplicate or sort values.
+#' @seealso [complete()] to expand list objects. [expand_grid()]
+#'   to input vectors rather than a data frame.
 #' @export
 #' @examples
-#' library(dplyr)
-#' # All possible combinations of vs & cyl, even those that aren't
-#' # present in the data
-#' expand(mtcars, vs, cyl)
-#'
-#' # Only combinations of vs and cyl that appear in the data
-#' expand(mtcars, nesting(vs, cyl))
-#'
-#' # Implicit missings ---------------------------------------------------------
-#' df <- tibble(
-#'   year   = c(2010, 2010, 2010, 2010, 2012, 2012, 2012),
-#'   qtr    = c(   1,    2,    3,    4,    1,    2,    3),
-#'   return = rnorm(7)
-#' )
-#' df %>% expand(year, qtr)
-#' df %>% expand(year = 2010:2012, qtr)
-#' df %>% expand(year = full_seq(year, 1), qtr)
-#' df %>% complete(year = full_seq(year, 1), qtr)
-#'
-#' # Nesting -------------------------------------------------------------------
-#' # Each person was given one of two treatments, repeated three times
-#' # But some of the replications haven't happened yet, so we have
-#' # incomplete data:
-#' experiment <- tibble(
-#'   name = rep(c("Alex", "Robert", "Sam"), c(3, 2, 1)),
-#'   trt  = rep(c("a", "b", "a"), c(3, 2, 1)),
-#'   rep = c(1, 2, 3, 1, 2, 1),
-#'   measurement_1 = runif(6),
-#'   measurement_2 = runif(6)
+#' fruits <- tibble(
+#'   type   = c("apple", "orange", "apple", "orange", "orange", "orange"),
+#'   year   = c(2010, 2010, 2012, 2010, 2010, 2012),
+#'   size  =  factor(
+#'     c("XS", "S",  "M", "S", "S", "M"),
+#'     levels = c("XS", "S", "M", "L")
+#'   ),
+#'   weights = rnorm(6, as.numeric(size) + 2)
 #' )
 #'
-#' # We can figure out the complete set of data with expand()
-#' # Each person only gets one treatment, so we nest name and trt together:
-#' all <- experiment %>% expand(nesting(name, trt), rep)
+#' # All possible combinations ---------------------------------------
+#' # Note that all defined, but not necessarily present, levels of the
+#' # factor variable `size` are retained.
+#' fruits %>% expand(type)
+#' fruits %>% expand(type, size)
+#' fruits %>% expand(type, size, year)
+#'
+#' # Only combinations that already appear in the data ---------------
+#' fruits %>% expand(nesting(type))
+#' fruits %>% expand(nesting(type, size))
+#' fruits %>% expand(nesting(type, size, year))
+#'
+#' # Other uses -------------------------------------------------------
+#' # Use with `full_seq()` to fill in values of continuous variables
+#' fruits %>% expand(type, size, full_seq(year, 1))
+#' fruits %>% expand(type, size, 2010:2012)
+#'
+#' # Use `anti_join()` to determine which observations are missing
+#' all <- fruits %>% expand(type, size, year)
 #' all
+#' all %>% dplyr::anti_join(fruits)
 #'
-#' # We can use anti_join to figure out which observations are missing
-#' all %>% anti_join(experiment)
-#'
-#' # And use right_join to add in the appropriate missing values to the
-#' # original data
-#' experiment %>% right_join(all)
-#' # Or use the complete() short-hand
-#' experiment %>% complete(nesting(name, trt), rep)
-#'
-#' # Generate all combinations with expand():
-#' formulas <- list(
-#'   formula1 = Sepal.Length ~ Sepal.Width,
-#'   formula2 = Sepal.Length ~ Sepal.Width + Petal.Width,
-#'   formula3 = Sepal.Length ~ Sepal.Width + Petal.Width + Petal.Length
-#' )
-#' data <- split(iris, iris$Species)
-#' crossing(formula = formulas, data)
+#' # Use with `right_join()` to fill in missing rows
+#' fruits %>% dplyr::right_join(all)
 expand <- function(data, ..., .name_repair = "check_unique") {
   UseMethod("expand")
 }