+++ R/rectangle.R
@@ -45,6 +45,9 @@
 #'   vector, by position with an integer vector, or with a combination of the
 #'   two with a list. See [purrr::pluck()] for details. The column names must
 #'   be unique. Existing columns with the same name are overwritten.
+#'
+#'   When plucking with a single string you can choose to omit the name;
+#'   `hoist(df, col, "x")` is short-hand for `hoist(df, col, x = "x")`.
 #' @param .simplify If `TRUE`, will attempt to simplify lists of length-1
 #'   vectors to an atomic vector
 #' @param .ptype Optionally, a named list of prototypes declaring the desired
@@ -78,7 +81,7 @@
 #'
 #' # Extract only specified components
 #' df %>% hoist(metadata,
-#'   species = "species",
+#'   "species",
 #'   first_film = list("films", 1L),
 #'   third_film = list("films", 3L)
 #' )
@@ -116,10 +119,7 @@ hoist <- function(.data, .col, ..., .remove = TRUE, .simplify = TRUE, .ptype = l
     abort("`col` must be a list-column")
   }
 
-  pluckers <- list2(...)
-  if (!is_named(pluckers)) {
-    stop("All elements of `...` must be named", call. = FALSE)
-  }
+  pluckers <- check_pluckers(...)
 
   if (vec_duplicate_any(names(pluckers))) {
     abort("The names of `...` must be unique")
@@ -156,6 +156,23 @@ hoist <- function(.data, .col, ..., .remove = TRUE, .simplify = TRUE, .ptype = l
   out[[.col]] <- x
 
   out
+}
+
+check_pluckers <- function(...) {
+  pluckers <- list2(...)
+
+  is_string <- vapply(pluckers, function(x) is.character(x) && length(x) == 1, logical(1))
+  auto_name <- names2(pluckers) == "" & is_string
+
+  if (any(auto_name)) {
+    names(pluckers)[auto_name] <- unlist(pluckers[auto_name])
+  }
+
+  if (!is_named(pluckers)) {
+    stop("All elements of `...` must be named", call. = FALSE)
+  }
+
+  pluckers
 }
 
 #' @export
@@ -352,7 +369,7 @@ simplify_col <- function(x, ptype, simplify = FALSE) {
   }
 
   # Ensure empty elements filled in with a single unspecified value
-  x[n == 0] <- list(unspecified(1))
+  x[n == 0] <- list(NA)
 
   tryCatch(
     vec_c(!!!x),